Updated logic & moved to services
